refactor: streamline CallHistoryList component and integrate action points management
- Replaced inline action points logic with a dedicated `useActionPointsManager` hook for better separation of concerns.
- Simplified the rendering of call items by utilizing a new `CallListItem` component, enhancing readability and maintainability.
- Updated the handling of selected call details and action points clearing logic for improved user experience.
